Missions only
The next piece of work on this particular text is to separate the
missions from the narrative. To find fields that can captrue all of
the informatiin. In days where two or more songs were played, the
dates will be separated into one entity per song. That also means
separating content where missions are separated into teams, e.g.
Red Team and Blue Team where this occurs in the timeline.