Create a JSON schema outline
This commit sees the first addition of new data in the form of NEW
HORIZONS which is very exciting. This is enabled by the nascent JSON
structure that is appearing. There is a better idea about how to
automate some of the remaining work now too with the mission log
looking more and more manageable.