Incremental changes to missions.log
This commit represents a series of edits to the already partially
normalized plain-text output of the NASA wake-up call PDF. These
edits are meant to delineate the information in a way that can be
parsed consistently with minimum coding. Folks wanting to look at
the original sources still have the initial commits to this
repository available. Work still needs to be done to understand what
to do with strange data, e.g. inline personal comments.