Add most of the unused data back into the JSON
There was some unused data for missions that had no songs. It has
now been added back into the JSON. That being date, the new field
MissionDates is not used consistently for all missions. We will
need to go back through the source and grab all of those at some
point as well.