QA Allinson and integrate new work
This commit re-introduces the nasawakeupcalls.json document as the
primary data source and integrates Matt's Discogs data with that.

Fields such as CAPCOM have been re-introduced to ensure that the
work remains as lossless as possible compared to Nasa's original
sources.

Minor QA was done on some of the information extracted from Discogs.

CSV output has been updated to ease QA and analysis for other folk.

Web-pages have been updated to enable searching on genre and style
and to link back to the Discogs database.